refactor: finalize the 3.2 package reorganization (C14)
- api-reference gains a Namespaces section (the twelve subpackages with
  scope and the 'from phonometry import aircraft as air' idiom); README
  quick-start notes the namespaces; CHANGELOG records the added namespaces,
  the deprecated flat module paths and the pickle note.
- plot_excitation moves from _plot/common.py to _plot/room.py and the
  facade imports it from there; _plot/common.py now holds only shared
  helpers and infrastructure.
- hearing/ and environmental/ package __init__ re-export the full public
  surface of their renamed modules (threshold, rating) so the pre-move
  package-path imports stay silent, as promised by the migration contract.
- SonarCloud cpd exclusion follows core.py to metrology/; llms.txt
  regenerated; .git-blame-ignore-revs lists the move commits and the local
  blame config registers it.

Full suite 2647 passed; conformance 194/194; golden baseline byte-stable;
import time unchanged vs main (1.09 s vs 1.17 s).
