+++ sonar-project.properties
@@ -19,7 +19,7 @@ sonar.python.coverage.reportPaths=coverage.xml
 sonar.exclusions=**/__pycache__/**, **/*.png, **/*.md
 sonar.test.inclusions=tests/**/test_*.py
 # Exclude overload-heavy type-stub files from duplication analysis
-sonar.cpd.exclusions=src/phonometry/__init__.py,src/phonometry/core.py
+sonar.cpd.exclusions=src/phonometry/__init__.py,src/phonometry/metrology/core.py
 
 # Increase authorized parameters for scientific APIs
 sonar.issue.ignore.multicriteria=S107
