+++ src/phonometry/_compat.py
@@ -48,7 +48,6 @@ _MOVED: dict[str, str] = {
     "phonometry.tonality": "phonometry.psychoacoustics.tonality",
     "phonometry.tonality_ecma": "phonometry.psychoacoustics.tonality_ecma",
     "phonometry.tone_audibility": "phonometry.psychoacoustics.tone_audibility",
-    "phonometry.hearing": "phonometry.hearing.threshold",
     "phonometry.noise_induced_hearing_loss": "phonometry.hearing.noise_induced_hearing_loss",
     "phonometry.occupational_exposure": "phonometry.hearing.occupational_exposure",
     "phonometry.sii": "phonometry.hearing.sii",
@@ -88,7 +87,6 @@ _MOVED: dict[str, str] = {
     "phonometry.multiple_shock_vibration": "phonometry.vibration.multiple_shock_vibration",
     "phonometry.transfer_stiffness": "phonometry.vibration.transfer_stiffness",
     "phonometry.air_absorption": "phonometry.environmental.air_absorption",
-    "phonometry.environmental": "phonometry.environmental.rating",
     "phonometry.environmental_measurement": "phonometry.environmental.measurement",
     "phonometry.impulse_prominence": "phonometry.environmental.impulse_prominence",
     "phonometry.outdoor_propagation": "phonometry.environmental.outdoor_propagation",
