+++ src/phonometry/_plot/common.py
@@ -525,85 +525,6 @@ def _time_axis(n: int, fs: int | None) -> tuple[np.ndarray, str]:
 
 
 
-def plot_excitation(
-    signal: "np.ndarray | Any",
-    fs: int,
-    *,
-    kind: str = "sweep",
-    ax: Axes | None = None,
-    **kwargs: Any,
-) -> Axes | np.ndarray:
-    """Plot an ISO 18233 excitation signal (sweep or MLS).
-
-    A documented helper for the raw arrays returned by
-    :func:`~phonometry.sweep_signal` and :func:`~phonometry.mls_signal`, which
-    stay plain :class:`numpy.ndarray` (they are meant for playback). For a
-    swept sine the waveform and its spectrogram are drawn; for an MLS the first
-    samples of the bipolar sequence and its (flat) magnitude spectrum.
-
-    :param signal: The excitation samples (1D array-like).
-    :param fs: Sample rate in Hz (for the time and frequency axes).
-    :param kind: ``"sweep"`` (default) or ``"mls"``.
-    :param ax: Existing axes for the top (time-domain) panel, or ``None`` for a
-        fresh two-panel figure.
-    :param kwargs: Forwarded to the time-domain ``plot`` call.
-    :return: The time-domain axes (``ax`` given) or the array of two axes.
-    """
-    x = np.asarray(signal, dtype=np.float64)
-    n = x.shape[-1]
-    if n == 0:
-        raise ValueError("excitation signal is empty; nothing to plot.")
-    t = np.arange(n) / float(fs)
-    color = kwargs.pop("color", _C_PRIMARY)
-
-    two_panel = ax is None
-    if two_panel:
-        axes = _new_axes_column(2, figsize=(8.0, 6.0))
-        ax_time = cast("Axes", axes[0])
-    else:
-        ax_time = cast("Axes", ax)
-
-    if kind == "mls":
-        show = min(n, 120)
-        ax_time.step(np.arange(show), x[:show], where="mid", color=color, **kwargs)
-        ax_time.set_xlabel("Sample")
-        ax_time.set_ylabel("Amplitude")
-        ax_time.set_ylim(-1.4, 1.4)
-        ax_time.set_title(
-            f"ISO 18233 MLS excitation (first {show} of {n} samples)"
-        )
-        ax_time.grid(True, alpha=0.3)
-        if not two_panel:
-            return ax_time
-        spec = np.abs(np.fft.rfft(x))
-        freqs = np.fft.rfftfreq(n, d=1.0 / fs)
-        ax_f = axes[1]
-        ac = spec[1:]
-        denom = float(np.median(ac)) if ac.size else 1.0
-        ax_f.semilogx(freqs[1:], 20.0 * np.log10(
-                      np.maximum(ac, 1e-10) / (denom if denom > 0.0 else 1.0)),
-                      color=_C_REFERENCE, lw=0.8)
-        ax_f.set_xlabel("Frequency [Hz]")
-        ax_f.set_ylabel("Magnitude [dB]")
-        ax_f.set_title("Magnitude spectrum (flat)")
-        ax_f.grid(True, which="both", alpha=0.3)
-        return axes
-
-    # Swept sine.
-    ax_time.plot(t, x, color=color, lw=0.6, **kwargs)
-    ax_time.set_xlabel("Time [s]")
-    ax_time.set_ylabel("Amplitude")
-    ax_time.set_title("ISO 18233 exponential sine sweep")
-    ax_time.grid(True, alpha=0.3)
-    if not two_panel:
-        return ax_time
-    ax_s = axes[1]
-    nperseg = min(n, max(256, min(2048, n // 16)))
-    ax_s.specgram(x, NFFT=nperseg, Fs=fs, noverlap=nperseg // 2, cmap="magma")
-    ax_s.set_xlabel("Time [s]")
-    ax_s.set_ylabel("Frequency [Hz]")
-    ax_s.set_title("Spectrogram (exponential frequency rise)")
-    return axes
 
 
 # ---------------------------------------------------------------------------
